fixed bug
Former-commit-id: d292c8e15e9178ef46eeea6ffc9b225c3efd6f3f

+++ src/gameengine/grid/ActorGrid.java
@@ -127,7 +127,7 @@ public class ActorGrid implements ReadableGrid, MasterGrid, ReadAndDamageGrid,
 	}
 	
 	@Override
-	public void controllerSpawnActor(Actor actor, int startX, int startY){
+	public void controllerSpawnActor(Actor actor, double startX, double startY){
 		addActor(actor, startX, startY);
 	}
