+++ src/gameengine/grid/interfaces/controllergrid/ControllableGrid.java
@@ -7,7 +7,7 @@ public interface ControllableGrid extends SteppableGrid{
 	
 	void move(int ID, double newX, double newY);
 	void removeActor(int ID);
-	void controllerSpawnActor(Actor actor, int startX, int startY);
+	void controllerSpawnActor(Actor actor, double startX, double startY);
 	Grid2D getLocationOf(int ID);
 	boolean isValidLoc(double x, double y);
